(조동욱) 결재 기능 구현중
1. 결재 리스트 구현
 - 분기에 따른 리스트 변경 작업 중
2. 결재 xml 파일 수정
3. 결재 추가 기능 구현중

+++ src/main/java/com/goodee/groupware/vo/ApprovalFile.java
@@ -9,7 +9,7 @@ public class ApprovalFile {
 	private String approvalFileOri;
 	private String approvalFileSave;
 	private String approvalFileType;
-	private int approvalFileSize;
+	private long approvalFileSize;
 	private String createdate;
 	private String updatedate;
 }
